# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for creating orders, managing bots, and viewing queue state.",
        "title": "Queue Management System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/bots": {
            "post": {
                "description": "Adds a bot that can process pending orders.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Add bot",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes the newest bot. Returns 404 if no bots exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Remove bot",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders": {
            "post": {
                "description": "Creates a new order. VIP orders are prioritized ahead of normal orders.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.orderResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/state": {
            "get": {
                "description": "Returns a JSON snapshot of queue state. WebSocket upgrades on this path stream updates.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "state"
                ],
                "summary": "Get state",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.stateResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.botResp": {
            "type": "object",
            "properties": {
                "current_order": {
                    "$ref": "#/definitions/api.orderResp"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "api.createOrderRequest": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "enum": [
                        "normal",
                        "vip"
                    ],
                    "example": "normal"
                }
            }
        },
        "api.orderResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.stateResp": {
            "type": "object",
            "properties": {
                "bots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.botResp"
                    }
                },
                "completed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.orderResp"
                    }
                },
                "pending": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.orderResp"
                    }
                }
            }
        }
    }
}